requirements.txt
flask
torch
torchvision
opencv-python
numpy
Pillow
google-generativeai
googletrans==4.0.0-rc1